+++ tests/helpers/models.ts
@@ -1,0 +1,11 @@
+import { CuiCachable } from "../../src/core/models/interfaces";
+
+export class CacheTestItem implements CuiCachable {
+    id: number;
+    constructor(id: number) {
+        this.id = id
+    }
+    refresh(): boolean {
+        return true;
+    }
+}