+++ changelog.md
@@ -1,4 +1,7 @@
 # bpd cUI Light changelog
+## [0.0.8] 2020-06-15
+* [Added] Method for async/interactive callbacks in element manager
+* [Added] Support for class **print** which reduces colors to black and white
 
 ## [0.0.7] 2020-06-14
 * [Fixed] Various fixes related to unit tests